chore(tools): use automation scripts from automation repo
