+++ tools/release/01_bump_dependency_crate_versions.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env -S deno run --allow-read --allow-write --allow-run=cargo
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
-import { DenoWorkspace } from "./helpers/mod.ts";
+import { DenoWorkspace } from "./deno_workspace.ts";
 
 const workspace = await DenoWorkspace.load();
 
@@ -8,4 +8,5 @@ for (const crate of workspace.getDependencyCrates()) {
   await crate.increment("minor");
 }
 
-await workspace.updateLockFile();
+// update the lock file
+await workspace.getCliCrate().cargoCheck();
