+++ tools/release/03_bump_cli_version.ts
@@ -1,19 +1,17 @@
 #!/usr/bin/env -S deno run --allow-read --allow-write --allow-run=cargo,git
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
-import {
-  DenoWorkspace,
-  formatGitLogForMarkdown,
-  getGitLogFromTag,
-} from "./helpers/mod.ts";
+import { DenoWorkspace } from "./deno_workspace.ts";
 
 const workspace = await DenoWorkspace.load();
+const repo = workspace.repo;
 
 const cliCrate = workspace.getCliCrate();
 const originalVersion = cliCrate.version;
 
 // increment the version
 await cliCrate.increment(getVersionIncrement());
-await workspace.updateLockFile();
+// update the lock file
+await cliCrate.cargoCheck();
 
 // output the Releases.md markdown text
 console.log(
@@ -34,11 +32,8 @@ function getVersionIncrement() {
 }
 
 async function getReleasesMdText() {
-  const gitLogOutput = await getGitLogFromTag(
-    DenoWorkspace.rootDirPath,
-    `v${originalVersion}`,
-  );
-  const formattedGitLog = formatGitLogForMarkdown(gitLogOutput);
+  const gitLog = await repo.getGitLogFromTag(`v${originalVersion}`);
+  const formattedGitLog = gitLog.formatForReleaseMarkdown();
   const formattedDate = getFormattedDate(new Date());
 
   return `### ${cliCrate.version} / ${formattedDate}\n\n` +
