+++ tools/release/deps.ts
@@ -1,0 +1,3 @@
+// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
+
+export * from "../../../../deno_automation/mod.ts";
