--- tools/release/helpers/mod.ts
@@ -1,10 +1,0 @@
-// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
-
-export * from "./cargo.ts";
-export * from "./crates_io.ts";
-export * from "./deno_workspace.ts";
-export {
-  formatGitLogForMarkdown,
-  getCratesPublishOrder,
-  getGitLogFromTag,
-} from "./helpers.ts";
